--- addons/ImGui/include/hk/gfx/ImGuiAllocator.h
@@ -1,13 +1,0 @@
-#pragma once
-
-#include "hk/types.h"
-
-namespace hk::gfx {
-    struct Allocator {
-        using Alloc = void* (*)(size, size);
-        using Free = void (*)(void*);
-
-        Alloc alloc = nullptr;
-        Free free = nullptr;
-    };
-}
